bot/requirements.txt
aiohttp==3.8.4
aioresponses==0.7.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
exceptiongroup==1.1.1
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
multidict==6.0.4
packaging==23.1
pluggy==1.0.0
pytest==7.3.1
pytest-asyncio==0.21.0
python-dotenv==1.0.0
python-telegram-bot==20.2
rfc3986==1.5.0
sniffio==1.3.0
tomli==2.0.1
yarl==1.9.2
